Collapse the flattening another time: BranchCommitList
The commit list contains the commit snippets
