+++ apps/desktop/src/components/v3/BranchCommitList.svelte
@@ -1,16 +1,42 @@
 <script lang="ts">
+	import CardOverlay from '$components/CardOverlay.svelte';
 	import CommitAction from '$components/CommitAction.svelte';
 	import CommitsAccordion from '$components/CommitsAccordion.svelte';
+	import Dropzone from '$components/Dropzone.svelte';
+	import LineOverlay from '$components/LineOverlay.svelte';
 	import ReduxResult from '$components/ReduxResult.svelte';
+	import CommitContextMenu, {
+		type CommitMenuContext
+	} from '$components/v3/CommitContextMenu.svelte';
+	import CommitGoesHere from '$components/v3/CommitGoesHere.svelte';
+	import CommitRow from '$components/v3/CommitRow.svelte';
+	import KebabButton from '$components/v3/KebabButton.svelte';
+	import { hasConflicts, isLocalAndRemoteCommit, isUpstreamCommit } from '$components/v3/lib';
+	import BaseBranchService from '$lib/baseBranch/baseBranchService.svelte';
+	import {
+		AmendCommitWithChangeDzHandler,
+		AmendCommitWithHunkDzHandler,
+		CommitDropData,
+		type DzCommitData,
+		SquashCommitDzHandler
+	} from '$lib/commits/dropHandler';
+	import { draggableCommitV3 } from '$lib/dragging/draggable';
+	import {
+		ReorderCommitDzFactory,
+		ReorderCommitDzHandler
+	} from '$lib/dragging/stackingReorderDropzoneManager';
+	import { DefaultForgeFactory } from '$lib/forge/forgeFactory.svelte';
 	import { StackService, type SeriesIntegrationStrategy } from '$lib/stacks/stackService.svelte';
 	import { combineResults } from '$lib/state/helpers';
+	import { UiState } from '$lib/state/uiState.svelte';
 	import { TestId } from '$lib/testing/testIds';
 	import { inject } from '@gitbutler/shared/context';
 	import Button from '@gitbutler/ui/Button.svelte';
 	import Modal from '@gitbutler/ui/Modal.svelte';
-	import type { Commit, UpstreamCommit } from '$lib/branches/v3';
-	import type { CommitKey } from '$lib/commits/commit';
-	import type { Snippet } from 'svelte';
+	import { getTimeAgo } from '@gitbutler/ui/utils/timeAgo';
+	import type { Commit } from '$lib/branches/v3';
+	import type { CommitStatusType } from '$lib/commits/commit';
+	import type { BranchDetails } from '$lib/stacks/stack';
 
 	const integrationStrategies = {
 		default: {
@@ -32,55 +58,66 @@
 	type IntegrationStrategy = keyof typeof integrationStrategies;
 
 	interface Props {
+		active: boolean;
 		projectId: string;
 		stackId: string;
 		branchName: string;
 		selectedCommitId?: string;
-		empty?: Snippet;
-		beforeLocalAndRemote?: Snippet;
-		upstreamTemplate?: Snippet<
-			[
-				{
-					commit: UpstreamCommit;
-					commitKey: CommitKey;
-					first: boolean;
-					lastCommit: boolean;
-				}
-			]
-		>;
-		localAndRemoteTemplate?: Snippet<
-			[
-				{
-					commit: Commit;
-					commitKey: CommitKey;
-					first: boolean;
-					last: boolean;
-					lastCommit: boolean;
-					ancestorMostConflicted: Commit | undefined;
-				}
-			]
-		>;
+		lastBranch: boolean;
+		branchDetails: BranchDetails;
+		stackingReorderDropzoneManager: ReorderCommitDzFactory;
+
+		handleUncommit: (commitId: string, branchName: string) => Promise<void>;
+		startEditingCommitMessage: (branchName: string, commitId: string) => void;
+		handleEditPatch: (args: {
+			commitId: string;
+			type: CommitStatusType;
+			hasConflicts: boolean;
+			isAncestorMostConflicted: boolean;
+		}) => void;
 	}
 
 	let {
+		active,
 		projectId,
 		stackId,
 		branchName,
-		empty,
-		beforeLocalAndRemote,
-		localAndRemoteTemplate,
-		upstreamTemplate
+		branchDetails,
+		selectedCommitId,
+		lastBranch,
+		stackingReorderDropzoneManager,
+		handleUncommit,
+		startEditingCommitMessage,
+		handleEditPatch
 	}: Props = $props();
 
-	const [stackService] = inject(StackService);
+	const [stackService, uiState, forge, baseBranchService] = inject(
+		StackService,
+		UiState,
+		DefaultForgeFactory,
+		BaseBranchService
+	);
 	const [integrateUpstreamCommits, upstreamIntegration] = stackService.integrateUpstreamCommits;
+
+	const projectState = $derived(uiState.project(projectId));
+	const drawer = $derived(projectState.drawerPage);
+	const isCommitting = $derived(drawer.current === 'new-commit');
+	const stackActive = $derived(stackId === projectState.stackId.current);
+	const stackState = $derived(uiState.stack(stackId));
+	const selection = $derived(stackState.selection.get());
+	const selectedBranchName = $derived(selection.current?.branchName);
 
 	const localAndRemoteCommits = $derived(stackService.commits(projectId, stackId, branchName));
 	const upstreamOnlyCommits = $derived(
 		stackService.upstreamCommits(projectId, stackId, branchName)
 	);
 
+	const baseBranchResponse = $derived(baseBranchService.baseBranch(projectId));
+	const base = $derived(baseBranchResponse.current.data);
+	const baseSha = $derived(base?.baseSha);
+
 	let confirmResetModal = $state<ReturnType<typeof Modal>>();
+	let commitMenuContext = $state<CommitMenuContext>();
 
 	async function integrate(strategy?: SeriesIntegrationStrategy): Promise<void> {
 		await integrateUpstreamCommits({
@@ -104,6 +141,13 @@
 			}
 		}
 		return undefined;
+	}
+
+	function handleCommitClick(commitId: string) {
+		const stackState = uiState.stack(stackId);
+		stackState.selection.set({ branchName, commitId });
+		uiState.project(projectId).drawerPage.set(undefined);
+		projectState.stackId.set(stackId);
 	}
 </script>
 
@@ -143,6 +187,14 @@
 	</Button>
 {/snippet}
 
+{#snippet commitReorderDz(dropzone: ReorderCommitDzHandler)}
+	<Dropzone handlers={[dropzone]}>
+		{#snippet overlay({ hovered, activated })}
+			<LineOverlay {hovered} {activated} />
+		{/snippet}
+	</Dropzone>
+{/snippet}
+
 <ReduxResult
 	{stackId}
 	{projectId}
@@ -152,8 +204,19 @@
 		{@const hasRemoteCommits = upstreamOnlyCommits.length > 0}
 		{@const hasCommits = localAndRemoteCommits.length > 0}
 		{@const ancestorMostConflicted = getAncestorMostConflicted(localAndRemoteCommits)}
-		{#if !hasCommits}
-			{@render empty?.()}
+		{#if !hasCommits && isCommitting}
+			<CommitGoesHere
+				selected={stackActive && branchName === selectedBranchName}
+				first
+				last
+				onclick={() => {
+					uiState.stack(stackId).selection.set({
+						branchName,
+						commitId: branchDetails.baseCommit
+					});
+					projectState.stackId.set(stackId);
+				}}
+			/>
 		{/if}
 		<div class="commit-list">
 			{#if hasRemoteCommits}
@@ -168,14 +231,35 @@
 						<span class="text-13 text-body text-semibold">Upstream commits</span>
 					{/snippet}
 
-					{#if upstreamTemplate}
-						{#each upstreamOnlyCommits as commit, i (commit.id)}
-							{@const first = i === 0}
-							{@const lastCommit = i === upstreamOnlyCommits.length - 1}
-							{@const commitKey = { stackId, branchName, commitId: commit.id, upstream: true }}
-							{@render upstreamTemplate({ commit, commitKey, first, lastCommit })}
-						{/each}
-					{/if}
+					{#each upstreamOnlyCommits as commit, i (commit.id)}
+						{@const first = i === 0}
+						{@const lastCommit = i === upstreamOnlyCommits.length - 1}
+						{@const selected =
+							stackActive && commit.id === selectedCommitId && branchName === selectedBranchName}
+						{@const commitId = commit.id}
+						{#if !isCommitting}
+							<CommitRow
+								type="Remote"
+								{stackId}
+								{commitId}
+								commitMessage={commit.message}
+								createdAt={commit.createdAt}
+								tooltip="Upstream"
+								{branchName}
+								{projectId}
+								{first}
+								lastCommit={lastCommit && localAndRemoteCommits.length === 0}
+								{selected}
+								{active}
+								onclick={() => {
+									uiState.stack(stackId).selection.set({ branchName, commitId, upstream: true });
+									uiState.project(projectId).drawerPage.set(undefined);
+									projectState.stackId.set(stackId);
+								}}
+								disableCommitActions={false}
+							/>
+						{/if}
+					{/each}
 
 					<CommitAction type="Remote" isLast={!hasCommits}>
 						{#snippet action()}
@@ -186,25 +270,174 @@
 				</CommitsAccordion>
 			{/if}
 
-			{#if localAndRemoteTemplate}
-				{@render beforeLocalAndRemote?.()}
-				{#each localAndRemoteCommits as commit, i (commit.id)}
-					{@const first = i === 0}
-					{@const last = i === localAndRemoteCommits.length - 1}
-					{@const commitKey = { stackId, branchName, commitId: commit.id, upstream: false }}
-					{@render localAndRemoteTemplate({
-						commit,
-						commitKey,
-						first,
-						last,
-						lastCommit: last,
-						ancestorMostConflicted
-					})}
-				{/each}
-			{/if}
+			{@render commitReorderDz(stackingReorderDropzoneManager.top(branchName))}
+			{#each localAndRemoteCommits as commit, i (commit.id)}
+				{@const first = i === 0}
+				{@const last = i === localAndRemoteCommits.length - 1}
+				{@const commitId = commit.id}
+				{@const selected =
+					stackActive && commit.id === selectedCommitId && branchName === selectedBranchName}
+				{#if isCommitting}
+					{@const nothingSelectedButFirst = selectedCommitId === undefined && first}
+					{@const selectedForCommit =
+						stackActive &&
+						(nothingSelectedButFirst || commit.id === selectedCommitId) &&
+						((first && selectedBranchName === undefined) || branchName === selectedBranchName)}
+					<!-- Only commits to the base can be `last`, see next `CommitGoesHere`. -->
+					<CommitGoesHere
+						selected={selectedForCommit}
+						{first}
+						last={false}
+						onclick={() => {
+							projectState.stackId.set(stackId);
+							uiState.stack(stackId).selection.set({ branchName, commitId });
+						}}
+					/>
+				{/if}
+				{@const dzCommit: DzCommitData = {
+					id: commit.id,
+					isRemote: isUpstreamCommit(commit),
+					isIntegrated: isLocalAndRemoteCommit(commit) && commit.state.type === 'Integrated',
+					hasConflicts: isLocalAndRemoteCommit(commit) && commit.hasConflicts,
+				}}
+				{@const amendHandler = new AmendCommitWithChangeDzHandler(
+					projectId,
+					stackService,
+					stackId,
+					dzCommit,
+					(newId) => uiState.stack(stackId).selection.set({ branchName, commitId: newId }),
+					uiState
+				)}
+				{@const squashHandler = new SquashCommitDzHandler({
+					stackService,
+					projectId,
+					stackId,
+					commit: dzCommit
+				})}
+				{@const hunkHandler = new AmendCommitWithHunkDzHandler({
+					stackService,
+					projectId,
+					stackId,
+					commit: dzCommit,
+					// TODO: Use correct value!
+					okWithForce: true,
+					uiState
+				})}
+				{@const tooltip = commit.state.type}
+				<Dropzone handlers={[amendHandler, squashHandler, hunkHandler]}>
+					{#snippet overlay({ hovered, activated, handler })}
+						{@const label =
+							handler instanceof AmendCommitWithChangeDzHandler ||
+							handler instanceof AmendCommitWithHunkDzHandler
+								? 'Amend'
+								: 'Squash'}
+						<CardOverlay {hovered} {activated} {label} />
+					{/snippet}
+					<div
+						use:draggableCommitV3={{
+							disabled: false,
+							label: commit.message.split('\n')[0],
+							sha: commit.id.slice(0, 7),
+							date: getTimeAgo(commit.createdAt),
+							authorImgUrl: undefined,
+							commitType: commit.state.type,
+							data: new CommitDropData(
+								stackId,
+								{
+									id: commitId,
+									isRemote: !!branchDetails.remoteTrackingBranch,
+									hasConflicts: isLocalAndRemoteCommit(commit) && commit.hasConflicts,
+									isIntegrated: isLocalAndRemoteCommit(commit) && commit.state.type === 'Integrated'
+								},
+								false,
+								branchName
+							),
+							viewportId: 'board-viewport'
+						}}
+					>
+						<CommitRow
+							commitId={commit.id}
+							commitMessage={commit.message}
+							type={commit.state.type}
+							hasConflicts={commit.state.type === 'LocalAndRemote' && commit.hasConflicts}
+							diverged={commit.state.type === 'LocalAndRemote' &&
+								commit.id !== commit.state.subject}
+							createdAt={commit.createdAt}
+							{stackId}
+							{branchName}
+							{projectId}
+							{first}
+							lastCommit={last}
+							{lastBranch}
+							{selected}
+							draggable
+							{tooltip}
+							{active}
+							isOpen={commit.id === commitMenuContext?.data.commitId}
+							onclick={() => handleCommitClick(commit.id)}
+							disableCommitActions={false}
+						>
+							{#snippet menu({ rightClickTrigger })}
+								{@const data = {
+									stackId,
+									commitId,
+									commitMessage: commit.message,
+									commitStatus: commit.state.type,
+									commitUrl: forge.current.commitUrl(commitId),
+									onUncommitClick: () => handleUncommit(commit.id, branchName),
+									onEditMessageClick: () => startEditingCommitMessage(branchName, commit.id),
+									onPatchEditClick: () =>
+										handleEditPatch({
+											commitId: commit.id,
+											type: commit.state.type,
+											hasConflicts: hasConflicts(commit),
+											isAncestorMostConflicted: ancestorMostConflicted?.id === commit.id
+										})
+								}}
+								<KebabButton
+									flat
+									contextElement={rightClickTrigger}
+									onclick={(element) => {
+										commitMenuContext = {
+											position: { element },
+											data
+										};
+									}}
+									oncontext={(coords) =>
+										(commitMenuContext = {
+											position: { coords },
+											data
+										})}
+									contextElementSelected={selected}
+									activated={commit.id === commitMenuContext?.data.commitId &&
+										!!commitMenuContext.position.element}
+								/>
+							{/snippet}
+						</CommitRow>
+					</div>
+				</Dropzone>
+				{@render commitReorderDz(stackingReorderDropzoneManager.belowCommit(branchName, commit.id))}
+				{#if isCommitting && last}
+					<CommitGoesHere
+						{first}
+						{last}
+						selected={stackActive &&
+							selectedCommitId === baseSha &&
+							branchName === selectedBranchName}
+						onclick={() => {
+							uiState.stack(stackId).selection.set({ branchName, commitId: baseSha });
+							projectState.stackId.set(stackId);
+						}}
+					/>
+				{/if}
+			{/each}
 		</div>
 	{/snippet}
 </ReduxResult>
+
+{#if commitMenuContext}
+	<CommitContextMenu {projectId} bind:context={commitMenuContext} />
+{/if}
 
 <style lang="postcss">
 	.commit-list {
