+++ apps/desktop/src/components/v3/BranchList.svelte
@@ -1,8 +1,5 @@
 <script lang="ts">
-	import CardOverlay from '$components/CardOverlay.svelte';
 	import ScrollableContainer from '$components/ConfigurableScrollableContainer.svelte';
-	import Dropzone from '$components/Dropzone.svelte';
-	import LineOverlay from '$components/LineOverlay.svelte';
 	import ReduxResult from '$components/ReduxResult.svelte';
 	import StackStickyButtons from '$components/StackStickyButtons.svelte';
 	import BranchCard from '$components/v3/BranchCard.svelte';
@@ -10,38 +7,13 @@
 	import BranchHeaderContextMenu, {
 		type BranchHeaderContextItem
 	} from '$components/v3/BranchHeaderContextMenu.svelte';
-	import CommitContextMenu, {
-		type CommitMenuContext
-	} from '$components/v3/CommitContextMenu.svelte';
-	import CommitGoesHere from '$components/v3/CommitGoesHere.svelte';
-	import CommitRow from '$components/v3/CommitRow.svelte';
 	import ConflictResolutionConfirmModal from '$components/v3/ConflictResolutionConfirmModal.svelte';
 	import KebabButton from '$components/v3/KebabButton.svelte';
 	import NewBranchModal from '$components/v3/NewBranchModal.svelte';
 	import PublishButton from '$components/v3/PublishButton.svelte';
 	import PushButton from '$components/v3/PushButton.svelte';
-
-	import {
-		getColorFromCommitState,
-		getIconFromCommitState,
-		hasConflicts,
-		isLocalAndRemoteCommit,
-		isUpstreamCommit
-	} from '$components/v3/lib';
-	import BaseBranchService from '$lib/baseBranch/baseBranchService.svelte';
-	import {
-		AmendCommitWithChangeDzHandler,
-		AmendCommitWithHunkDzHandler,
-		CommitDropData,
-		SquashCommitDzHandler,
-		type DzCommitData
-	} from '$lib/commits/dropHandler';
-	import { draggableCommitV3 } from '$lib/dragging/draggable';
-	import {
-		ReorderCommitDzHandler,
-		StackingReorderDropzoneManagerFactory
-	} from '$lib/dragging/stackingReorderDropzoneManager';
-	import { DefaultForgeFactory } from '$lib/forge/forgeFactory.svelte';
+	import { getColorFromCommitState, getIconFromCommitState } from '$components/v3/lib';
+	import { StackingReorderDropzoneManagerFactory } from '$lib/dragging/stackingReorderDropzoneManager';
 	import { ModeService } from '$lib/mode/modeService';
 	import { StackService } from '$lib/stacks/stackService.svelte';
 	import { combineResults } from '$lib/state/helpers';
@@ -49,7 +21,6 @@
 	import { getContext, inject } from '@gitbutler/shared/context';
 	import Button from '@gitbutler/ui/Button.svelte';
 	import Modal from '@gitbutler/ui/Modal.svelte';
-	import { getTimeAgo } from '@gitbutler/ui/utils/timeAgo';
 	import type { CommitStatusType } from '$lib/commits/commit';
 
 	type Props = {
@@ -60,19 +31,9 @@
 	};
 
 	const { isVerticalMode, projectId, stackId, active }: Props = $props();
-	const [stackService, uiState, baseBranchService, forge, modeService] = inject(
-		StackService,
-		UiState,
-		BaseBranchService,
-		DefaultForgeFactory,
-		ModeService
-	);
+	const [stackService, uiState, modeService] = inject(StackService, UiState, ModeService);
 
 	const branchesResult = $derived(stackService.branches(projectId, stackId));
-
-	const baseBranchResponse = $derived(baseBranchService.baseBranch(projectId));
-	const base = $derived(baseBranchResponse.current.data);
-	const baseSha = $derived(base?.baseSha);
 
 	const projectState = $derived(uiState.project(projectId));
 	const drawer = $derived(projectState.drawerPage);
@@ -82,7 +43,6 @@
 	const stackState = $derived(uiState.stack(stackId));
 	const selection = $derived(stackState.selection.get());
 	const selectedCommitId = $derived(selection.current?.commitId);
-	const selectedBranchName = $derived(selection.current?.branchName);
 
 	let newBranchModal = $state<ReturnType<typeof NewBranchModal>>();
 
@@ -96,7 +56,7 @@
 		if (branchName) stackState.selection.set({ branchName, commitId: undefined });
 	}
 
-	function startCommitMessageEdition(branchName: string, commitId: string) {
+	function startEditingCommitMessage(branchName: string, commitId: string) {
 		stackState.selection.set({ branchName, commitId });
 		projectState.drawerPage.set(undefined);
 		projectState.editingCommitMessage.set(true);
@@ -115,8 +75,21 @@
 		modeService!.enterEditMode(args.commitId, stackId);
 	}
 
+	function handleBranchClick(branchName: string, headCommit: string) {
+		uiState.project(projectId).stackId.set(stackId);
+		if (isCommitting) {
+			uiState.stack(stackId).selection.set({
+				branchName,
+				commitId: headCommit
+			});
+			projectState.stackId.set(stackId);
+		} else {
+			uiState.stack(stackId).selection.set({ branchName });
+			uiState.project(projectId).drawerPage.set('branch');
+		}
+	}
+
 	let headerMenuContext = $state<BranchHeaderContextItem>();
-	let commitMenuContext = $state<CommitMenuContext>();
 
 	const stackingReorderDropzoneManagerFactory = getContext(StackingReorderDropzoneManagerFactory);
 </script>
@@ -202,17 +175,7 @@
 									trackingBranch={branch.remoteTrackingBranch ?? undefined}
 									readonly={!!branch.remoteTrackingBranch}
 									onclick={() => {
-										uiState.project(projectId).stackId.set(stackId);
-										if (isCommitting) {
-											uiState.stack(stackId).selection.set({
-												branchName,
-												commitId: headCommit
-											});
-											projectState.stackId.set(stackId);
-										} else {
-											uiState.stack(stackId).selection.set({ branchName });
-											uiState.project(projectId).drawerPage.set('branch');
-										}
+										handleBranchClick(branchName, headCommit);
 									}}
 								>
 									{#snippet menu({ rightClickTrigger })}
@@ -232,248 +195,19 @@
 										/>
 									{/snippet}
 									{#snippet commitList()}
-										{#snippet commitReorderDz(dropzone: ReorderCommitDzHandler)}
-											<Dropzone handlers={[dropzone]}>
-												{#snippet overlay({ hovered, activated })}
-													<LineOverlay {hovered} {activated} />
-												{/snippet}
-											</Dropzone>
-										{/snippet}
-										<BranchCommitList {projectId} {stackId} {branchName} {selectedCommitId}>
-											{#snippet empty()}
-												{#if isCommitting}
-													<CommitGoesHere
-														selected={stackActive && branchName === selectedBranchName}
-														first
-														last
-														onclick={() => {
-															uiState.stack(stackId).selection.set({
-																branchName,
-																commitId: branchDetails.baseCommit
-															});
-															projectState.stackId.set(stackId);
-														}}
-													/>
-												{/if}
-											{/snippet}
-											{#snippet upstreamTemplate({ commit, first, lastCommit })}
-												{@const selected =
-													stackActive &&
-													commit.id === selectedCommitId &&
-													branchName === selectedBranchName}
-												{@const commitId = commit.id}
-												{#if !isCommitting}
-													<CommitRow
-														type="Remote"
-														{stackId}
-														{commitId}
-														commitMessage={commit.message}
-														createdAt={commit.createdAt}
-														tooltip="Upstream"
-														{branchName}
-														{projectId}
-														{first}
-														lastCommit={lastCommit && localAndRemoteCommits.length === 0}
-														{selected}
-														{active}
-														onclick={() => {
-															uiState
-																.stack(stackId)
-																.selection.set({ branchName, commitId, upstream: true });
-															uiState.project(projectId).drawerPage.set(undefined);
-															projectState.stackId.set(stackId);
-														}}
-														disableCommitActions={false}
-													/>
-												{/if}
-											{/snippet}
-											{#snippet beforeLocalAndRemote()}
-												{@render commitReorderDz(stackingReorderDropzoneManager.top(branch.name))}
-											{/snippet}
-											{#snippet localAndRemoteTemplate({
-												commit,
-												first,
-												last,
-												lastCommit,
-												ancestorMostConflicted
-											})}
-												{@const commitId = commit.id}
-												{@const selected =
-													stackActive &&
-													commit.id === selectedCommitId &&
-													branchName === selectedBranchName}
-												{#if isCommitting}
-													{@const nothingSelectedButFirst = selectedCommitId === undefined && first}
-													{@const selectedForCommit =
-														stackActive &&
-														(nothingSelectedButFirst || commit.id === selectedCommitId) &&
-														((first && selectedBranchName === undefined) ||
-															branchName === selectedBranchName)}
-													<!-- Only commits to the base can be `last`, see next `CommitGoesHere`. -->
-													<CommitGoesHere
-														selected={selectedForCommit}
-														{first}
-														last={false}
-														onclick={() => {
-															projectState.stackId.set(stackId);
-															uiState.stack(stackId).selection.set({ branchName, commitId });
-														}}
-													/>
-												{/if}
-												{@const dzCommit: DzCommitData = {
-													id: commit.id,
-													isRemote: isUpstreamCommit(commit),
-													isIntegrated: isLocalAndRemoteCommit(commit) && commit.state.type === 'Integrated',
-													hasConflicts: isLocalAndRemoteCommit(commit) && commit.hasConflicts,
-												}}
-												{@const amendHandler = new AmendCommitWithChangeDzHandler(
-													projectId,
-													stackService,
-													stackId,
-													dzCommit,
-													(newId) =>
-														uiState.stack(stackId).selection.set({ branchName, commitId: newId }),
-													uiState
-												)}
-												{@const squashHandler = new SquashCommitDzHandler({
-													stackService,
-													projectId,
-													stackId,
-													commit: dzCommit
-												})}
-												{@const hunkHandler = new AmendCommitWithHunkDzHandler({
-													stackService,
-													projectId,
-													stackId,
-													commit: dzCommit,
-													// TODO: Use correct value!
-													okWithForce: true,
-													uiState
-												})}
-												{@const tooltip = commit.state.type}
-												<Dropzone handlers={[amendHandler, squashHandler, hunkHandler]}>
-													{#snippet overlay({ hovered, activated, handler })}
-														{@const label =
-															handler instanceof AmendCommitWithChangeDzHandler ||
-															handler instanceof AmendCommitWithHunkDzHandler
-																? 'Amend'
-																: 'Squash'}
-														<CardOverlay {hovered} {activated} {label} />
-													{/snippet}
-													<div
-														use:draggableCommitV3={{
-															disabled: false,
-															label: commit.message.split('\n')[0],
-															sha: commit.id.slice(0, 7),
-															date: getTimeAgo(commit.createdAt),
-															authorImgUrl: undefined,
-															commitType: commit.state.type,
-															data: new CommitDropData(
-																stackId,
-																{
-																	id: commitId,
-																	isRemote: !!branchDetails.remoteTrackingBranch,
-																	hasConflicts:
-																		isLocalAndRemoteCommit(commit) && commit.hasConflicts,
-																	isIntegrated:
-																		isLocalAndRemoteCommit(commit) &&
-																		commit.state.type === 'Integrated'
-																},
-																false,
-																branchName
-															),
-															viewportId: 'board-viewport'
-														}}
-													>
-														<CommitRow
-															commitId={commit.id}
-															commitMessage={commit.message}
-															type={commit.state.type}
-															hasConflicts={commit.state.type === 'LocalAndRemote' &&
-																commit.hasConflicts}
-															diverged={commit.state.type === 'LocalAndRemote' &&
-																commit.id !== commit.state.subject}
-															createdAt={commit.createdAt}
-															{stackId}
-															{branchName}
-															{projectId}
-															{first}
-															{lastCommit}
-															{lastBranch}
-															{selected}
-															draggable
-															{tooltip}
-															{active}
-															isOpen={commit.id === commitMenuContext?.data.commitId}
-															onclick={() => {
-																const stackState = uiState.stack(stackId);
-																stackState.selection.set({ branchName, commitId });
-																uiState.project(projectId).drawerPage.set(undefined);
-																projectState.stackId.set(stackId);
-															}}
-															disableCommitActions={false}
-														>
-															{#snippet menu({ rightClickTrigger })}
-																{@const data = {
-																	stackId,
-																	commitId,
-																	commitMessage: commit.message,
-																	commitStatus: commit.state.type,
-																	commitUrl: forge.current.commitUrl(commitId),
-																	onUncommitClick: () => handleUncommit(commit.id, branch.name),
-																	onEditMessageClick: () =>
-																		startCommitMessageEdition(branchName, commit.id),
-																	onPatchEditClick: () =>
-																		handleEditPatch({
-																			commitId: commit.id,
-																			type: commit.state.type,
-																			hasConflicts: hasConflicts(commit),
-																			isAncestorMostConflicted:
-																				ancestorMostConflicted?.id === commit.id
-																		})
-																}}
-																<KebabButton
-																	flat
-																	contextElement={rightClickTrigger}
-																	onclick={(element) => {
-																		commitMenuContext = {
-																			position: { element },
-																			data
-																		};
-																	}}
-																	oncontext={(coords) =>
-																		(commitMenuContext = {
-																			position: { coords },
-																			data
-																		})}
-																	contextElementSelected={selected}
-																	activated={commit.id === commitMenuContext?.data.commitId &&
-																		!!commitMenuContext.position.element}
-																/>
-															{/snippet}
-														</CommitRow>
-													</div>
-												</Dropzone>
-												{@render commitReorderDz(
-													stackingReorderDropzoneManager.belowCommit(branch.name, commit.id)
-												)}
-												{#if isCommitting && last}
-													<CommitGoesHere
-														{first}
-														{last}
-														selected={stackActive &&
-															selectedCommitId === baseSha &&
-															branchName === selectedBranchName}
-														onclick={() => {
-															uiState
-																.stack(stackId)
-																.selection.set({ branchName, commitId: baseSha });
-															projectState.stackId.set(stackId);
-														}}
-													/>
-												{/if}
-											{/snippet}
-										</BranchCommitList>
+										<BranchCommitList
+											{lastBranch}
+											{active}
+											{projectId}
+											{stackId}
+											{branchName}
+											{selectedCommitId}
+											{branchDetails}
+											{stackingReorderDropzoneManager}
+											{handleUncommit}
+											{startEditingCommitMessage}
+											{handleEditPatch}
+										/>
 									{/snippet}
 								</BranchCard>
 							{/snippet}
@@ -520,10 +254,6 @@
 
 {#if headerMenuContext}
 	<BranchHeaderContextMenu {projectId} {stackId} bind:context={headerMenuContext} />
-{/if}
-
-{#if commitMenuContext}
-	<CommitContextMenu {projectId} bind:context={commitMenuContext} />
 {/if}
 
 <style lang="postcss">
